requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.3.5
faiss-cpu==1.9.0.post1
numpy>=1.21.0
python-dotenv==1.0.0
gunicorn==21.2.0
